Consolidate function logic to a single file
